fix bugs in `cmd.new.bash`

+++ cmd/new/bash.bash
@@ -5,29 +5,19 @@ here=`cd $(dirname ${BASH_SOURCE[0]}) && pwd`
 env=`cat "${1}/env"`
 shift
 
+## Get cmd name
 name="${1}"
 if [ -z "${name}" ]; then
 	echo '[:(] arg "cmd-name" is empty' >&2
 	exit 1
 fi
-
 name=`echo "${name}" | tr '.' '/'`
-cmd_path=`get_path_under_pwd "${env}" "${name}"`
 
+## Get pwd
 curr_dir=`get_path_under_pwd "${env}" 'dummy'`
 curr_dir=`dirname "${curr_dir}"`
 
-if [ -e "${cmd_path}.bash" ]; then
-	echo "[:(] cmd script file '${cmd_path}.bash' exists, exited" >&2
-	exit 1
-fi
-if [ -e "${cmd_path}.bash.ticat" ]; then
-	echo "[:(] cmd meta file '${cmd_path}.bash.ticat' exists, exited" >&2
-	exit 1
-fi
-
-mkdir -p `dirname "${cmd_path}"`
-
+## Get repo root, make sure it's a repo dir
 repo_root=`find_repo_root_dir "${curr_dir}"`
 if [ -z "${repo_root}" ]; then
 	repo_root="${curr_dir}"
@@ -38,9 +28,11 @@ if [ -z "${repo_root}" ]; then
 	git init 2>&1 | awk '{print "     [git] "$0}'
 	echo
 fi
+repo_root=`abs_path "${repo_root}"`
+cd "${repo_root}"
 
+## Make sure bash helper lib is downloaded
 if [ ! -e "${repo_root}/helper/bash.helper" ]; then
-	cd "${repo_root}"
 	mkdir -p "helper"
 	echo "[:-] prepare to do 'git submodule add ...' to download bash helper libs"
 	echo '     ---'
@@ -50,11 +42,28 @@ if [ ! -e "${repo_root}/helper/bash.helper" ]; then
 	echo
 fi
 
-rel_path=`rel_path_to_repo_root "${curr_dir}" "${repo_root}"`
+## Make sure the old script/meta file not exists
+cmd_path="${repo_root}/${name}"
+if [ -e "${cmd_path}.bash" ]; then
+	echo "[:(] cmd script file '${cmd_path}.bash' exists, exited" >&2
+	exit 1
+fi
+if [ -e "${cmd_path}.bash.ticat" ]; then
+	echo "[:(] cmd meta file '${cmd_path}.bash.ticat' exists, exited" >&2
+	exit 1
+fi
+
+## Create the dir
+cmd_dir=`dirname "${cmd_path}"`
+mkdir -p "${cmd_dir}"
+
+## Create the script file
+rel_path=`rel_path_to_repo_root "${cmd_dir}" "${repo_root}"`
 echo 'set -euo pipefail' >> "${cmd_path}.bash"
 echo '. "`cd $(dirname ${BASH_SOURCE[0]}) && pwd`/'${rel_path}'/helper/helper.bash"' >> "${cmd_path}.bash"
 cat "${here}/templates/bash/simple.bash" >> "${cmd_path}.bash"
 echo "[:)] '${cmd_path}.bash' (cmd script file) created"
 
+## Create the meta file
 cp -f "${here}/templates/bash/simple.bash._ticat" "${cmd_path}.bash.ticat"
 echo "[:)] '${cmd_path}.bash.ticat' (cmd meta file) created"
